requirements.txt
# =============================================================================
# SIGE — dependências (arquivo único)
# =============================================================================
# Instalação: pip install -r requirements.txt
# O pip lê este arquivo de cima para baixo; as seções são só organização visual.
# =============================================================================

# -----------------------------------------------------------------------------
# Execução — framework, API, integrações usadas em produção / runserver
# -----------------------------------------------------------------------------
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-widget-tweaks==1.5.1
djangorestframework==3.17.1
mysqlclient==2.2.8
pillow==12.2.0
python-decouple==3.8
python-dotenv==1.2.2
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3

# -----------------------------------------------------------------------------
# Desenvolvimento & CI — formatação, lint, tipagem, testes
# -----------------------------------------------------------------------------
black==26.3.1
coverage==7.13.5
django-stubs==6.0.2
django-stubs-ext==6.0.2
flake8==7.3.0
isort==8.0.1
mypy==1.20.1
pylint==4.0.5
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
types-PyYAML==6.0.12.20250915

# -----------------------------------------------------------------------------
# Nota: demais pacotes (ex.: pycodestyle, pyflakes, astroid) vêm como
# dependências dos itens acima — não é preciso listar tudo manualmente.
# -----------------------------------------------------------------------------
